LearnLooper.jl

LearnLooper.learn_loop — Method

learn_loop(input, spans; num_repetitions, iteration_mode) -> nothing

#TODO-future: real docstring!

clarify difference between 0 and 1 repetitions; consider "num_playbacks" or similar

note that non-contiguous spans may result in a click in cumulative iteration mode

Non-exported functions and types

LearnLooper.pause — Method

pause(input) -> nothing

Computer pauses for duration it would take to play input.

LearnLooper.play — Method

play(input; volume_scale=1) -> nothing

Computer plays span of input. Currently only supports Mac playback!